refactor(smart-mic-pwa): extract component helpers and polish effects
Move pure helpers (TABS, formatElapsed, clearCachesAndReload, visible count
constants, translation direction helpers) out of their components into
colocated .helpers.ts files to keep components focused on rendering.

Also clean up the enter-button repeat timer on unmount, fix rec-area touch
effect deps, and convert the smartmic reducer to arrow function style.
