+++ src/smartmic/components/transcription-zone.helpers.ts
@@ -1,0 +1,2 @@
+export const COPY_FEEDBACK_MS = 2000;
+export const REMOTE_VISIBLE_COUNT = 3;
